Release 1.0
Major changes include proper testing of logging and it's masking routines. It is also now designed to 'never starve' connections, if one is not available within 2 seconds (by default), it will create on the calling thread rather than throw an exception.

Also configured to use the Java Cleaner API and now requires Java 9 at the minimum.